docs(validation): add Javadoc comments to ValidationResult
